fix: Ensure transpiled file is written to disk
Previously, nutra-coverage and nutra-babel did not account for the
possibility of a second transpiler. This resulted in the two
plugins being tighly coupled in the sense that nutra-coverage was
the one writing the nutra-babel file to disk. This commit decouples
nutra-coverage by having it receive filename from the previous
plugin and also resolve the sourcemaps based on a reconstruction
chain. Then write a final conherent version to disk.